requirements.txt
scapy
pyyaml
ruamel.yaml
utils
argparse
yamale 